implement "forwarded" argument for blacklist
Makes it possible to deny forwarding only. I want my router box to be
able to interact with some of the hosts in a blacklist (e.g. public DNS
resolvers), but I do not want my LAN hosts to be able to talk directly
to such hosts. So I can specify the "forward"/"forwarded" argument and
skip creating any INPUT/OUTPUT chain rules.

Signed-off-by: Steven Noonan <steven@uplinklabs.net>